Remove redundant `FrameBuffer()` construction
It seems `this->frameBuffer` is already initialized before entering
the body of `SSD1306()`. Therefore there is no need to construct another
`FrameBuffer`. In fact, doing so can cause memory issues when
`~FrameBuffer()` is called upon leaving `SSD1306()`, with the freed
buffer in the deconstructed `FrameBuffer` being written to by the
constructed `SSD1306` when used.